changed fallback macro to have the same return type as rint().
2008-05-21  Sven Neumann  <sven@gimp.org>

	* babl/babl-internal.h (rint): changed fallback macro to have the
	same return type as rint().

svn path=/trunk/; revision=314

+++ babl/babl-internal.h
@@ -54,7 +54,7 @@
 
 /* fallback to floor function when rint is not around */
 #ifndef HAVE_RINT
-# define rint(f)  ((long) floor (((double) f) + 0.5))
+# define rint(f)  (floor (((double) (f)) + 0.5))
 #endif
